+++ cogs/admin.py
@@ -19,7 +19,7 @@ class AdminCog(commands.Cog, name="Admin"):
 
     Arguments:
     ---
-        `bot` `discord.commands.Bot` -- The bot class that deals with all the commands.
+        bot {discord.commands.Bot} -- The bot
     """
     def __init__(self, bot):
         self.bot = bot