+++ cogs/misc.py
@@ -20,7 +20,7 @@ class MiscCog(commands.Cog, name="Misc"):
 
     Arguments:
     ---
-        `bot` `discord.commands.Bot` -- The bot class that deals with all the commands.
+        bot {discord.commands.Bot} -- The bot
     """
     def __init__(self, bot):
         self.bot = bot