+++ cogs/regions.py
@@ -17,7 +17,7 @@ class RegionCog(commands.Cog, name="Regions"):
 
     Arguments:
     ---
-        `bot` `discord.commands.Bot` -- The bot class that deals with all the commands.
+        bot {discord.commands.Bot} -- The bot
     """
     def __init__(self, bot):
         self.bot = bot